backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.30
pydantic==2.7.0
pydantic-settings==2.2.1
httpx==0.27.0
python-dotenv==1.0.1
tenacity==8.2.3
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
